TST: Don't put np.nan into an integer array.

+++ dask/array/tests/test_reductions.py
@@ -542,7 +542,7 @@ def test_array_cumreduction_axis(func, use_nan, axis, method):
     da_func = getattr(da, func)
 
     s = (10, 11, 12)
-    a = np.arange(np.prod(s)).reshape(s)
+    a = np.arange(np.prod(s), dtype=float).reshape(s)
     if use_nan:
         a[1] = np.nan
     d = da.from_array(a, chunks=(4, 5, 6))
